adicionando entidade InsuranceCompany para nao precisar colocar nome em todos seguros e atualizando para quando mudar a marca do modelo, mudar tbm em todos veiculos

+++ src/main/java/com/nozama/aluguel_veiculos/dto/VehicleRequest.java
@@ -31,7 +31,7 @@ public record VehicleRequest(
         Integer idCategoria,
 
         @NotNull(message = "Informe o seguro do veículo.")
-        Integer idSeguro,
+        Long idSeguro,
 
         @NotNull(message = "Informe o valor da diaria do veículo")
         BigDecimal valorDiario
@@ -43,7 +43,7 @@ public record VehicleRequest(
             String status,
             String descricao,
             Integer idCategoria,
-            Integer idSeguro,
+            Long idSeguro,
             Integer ano,
             BigDecimal valorDiario
     ) {
