+++ src/main/java/com/nozama/aluguel_veiculos/repository/InsuranceRepository.java
@@ -3,5 +3,5 @@ package com.nozama.aluguel_veiculos.repository;
 import com.nozama.aluguel_veiculos.domain.Insurance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InsuranceRepository extends JpaRepository<Insurance, Integer> {
+public interface InsuranceRepository extends JpaRepository<Insurance, Long> {
 }
